+++ tests/test_pyinstaller_utils.py
@@ -86,10 +86,6 @@ def test_pyinstaller_datas():
                 "kimi_cli/tools/file",
             ),
             (
-                "src/kimi_cli/tools/file/patch.md",
-                "kimi_cli/tools/file",
-            ),
-            (
                 "src/kimi_cli/tools/file/read.md",
                 "kimi_cli/tools/file",
             ),
@@ -136,7 +132,6 @@ def test_pyinstaller_hiddenimports():
             "kimi_cli.tools.file",
             "kimi_cli.tools.file.glob",
             "kimi_cli.tools.file.grep",
-            "kimi_cli.tools.file.patch",
             "kimi_cli.tools.file.read",
             "kimi_cli.tools.file.replace",
             "kimi_cli.tools.file.write",